Get DOM working
Added event listeners to the squares created by the game board creation
and add logic that places X's and O's to the visible game board. Game
still runs primarily from the console. Likely going to stop the project
here to continue learning the next topics as I feel I have achieved the
purpose of the assignment with factory functions and methods and am
interested in continuing my learning.